+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add prometheus metrics about dex-app registrations on identity providers
+
 ## [0.1.3] - 2022-12-07
 
 ### Changed